[SC-383] Re-enable updating R and R-Studio
This re-enables commit af67094a2c5993d352c9b7c1908814695f754654

That commit was reverted in commit 409e294 to get to the last passing
build state to make it easier to fix the recent CI build breakage.
Now that the build breakage is fixed we can now re-enable the
rstudio update commit.